requirements.txt
# analyze_requirements.txt
boto3
python-docx
PyMuPDF
requests 